don't interpret commands for the current program
This should avoid some confusion with either solution to a dilemma. If I show
the current program interpreted as if it was being run from the beginning in
the current mode, then this may look like different instructions were executed
than actually were. On the other hand if I show real history and future
extrapolation, it may look like the program will execute different things from
the beginning in the current mode. So just leave any interpretation of the
current program to the user.