cleaner clue highlighting code, fixes glitches
Treating the highlight logic as something to be driven externally rather than
having it dig into the depths of the game state to figure out what is going on.
The clues display doesn't have to be a sprite as I have here, but this approach
may lend itself to animation later on so I figure it makes sense.