remove level 2a from rotation
Also fix its noncontiguous program regions. It now has too many regions for the
a-f lettering that I have available, so that needs to be fixed before it can be
reinstituted (either add g or consolidate).

Generally I just have been very uncomfortable about this level, it is trickier
and more confusing than it needs to be, at least the back half.